removed export from index.js
