+++ shoppingCart/index.html
@@ -105,9 +105,7 @@
       </div>
     </div>
   </div>
-  <script type="module" src="app.js"></script>
-  <script type="module" src="./js/getterMethods.js"></script>
-  <script type="module" src="./js/utilFunctions.js"></script>
+  <script type="module" src="index.js"></script>
 </body>
 
 </html>
